refactor: clean-up types
