+++ src/types/types.ts
@@ -1,18 +1,22 @@
 // Enigma model
 
-import { PivotData } from "../pivot-table/handle-data";
-import { NxPageArea, NxPivotPage } from "./QIX";
+import { NxPageArea, NxPivotDimensionCell, NxPivotPage, NxPivotValuePoint } from "./QIX";
+
+export enum TYPE {
+  LABEL = 'LABEL',
+  DIMENSION = 'DIMENSION',
+  MEASURE = 'MEASURE',
+  EMPTY = 'EMPTY',
+};
 
 type ExpandFn = (qHyperCubeDef: string, rownNumber: number, column: number, option: boolean) => void;
-type GetHyperCubePivotDataParam = { qPath: string, qPages: Array<NxPageArea> };
-type GetHyperCubePivotData = (param: GetHyperCubePivotDataParam) => Promise<NxPivotPage[]>
 
 export interface Model {
   expandLeft: ExpandFn
   collapseLeft: ExpandFn
   expandTop: ExpandFn
   collapseTop: ExpandFn
-  getHyperCubePivotData: GetHyperCubePivotData
+  getHyperCubePivotData: (param: { qPath: string, qPages: Array<NxPageArea> }) => Promise<NxPivotPage[]>
 }
 
 export interface Rect {
@@ -20,64 +24,25 @@ export interface Rect {
   height: number;
 }
 
-export interface Item {
-  qElemNo: number;
-  column: number;
-  qText: string;
-  qNum: string;
-}
-
-export interface ColumnItem extends Item {
-  qCanExpand: boolean;
-  qCanCollapse: boolean;
-  column: number;
-}
-
-export interface Column {
-  headers: Array<Dimension>;
-}
-
-export interface PivotColumns {
-  items: Array<ColumnItem>;
-  headers: Array<Dimension>;
-}
-
-export interface PivotTableData {
-  leftColumns: PivotColumns;
-  topColumns: PivotColumns;
-  rows: Array<Array<Item>>;
-  rootIndex: Array<number>
-}
-
-// Object layout
-export interface ColumnNode {
-  qCanExpand: boolean;
-  qCanCollapse: boolean;
-  qElemNo: number;
-  qSubNodes: Array<ColumnNode>;
-  qText: string;
-  qType: string;
-  qUp: number;
-  qValue: string;
-}
-
-export interface DataPage {
-  qTop: Array<ColumnNode>;
-  qLeft: Array<ColumnNode>;
-  qData: Array<Array<Item>>;
-}
-
-export interface Dimension {
-  cId: string;
-  qFallbackTitle: string;
-}
-
 export interface ItemData {
   pivotData: PivotData;
   model: Model;
 }
 
-export type QPivotDataPages = Array<DataPage>;
-export type QDimensionInfo = Array<Dimension>;
+export interface Cell {
+  key: number | string;
+  type: string;
+  value: CellValue;
+}
 
-export type ElementRef = { current: HTMLElement } | undefined;
+export type CellValue = NxPivotValuePoint | NxPivotDimensionCell | string | null | undefined;
+
+export type Matrix = Array<Cell[]>;
+
+export interface PivotData {
+  matrix: Matrix;
+  leftMatrix: Matrix,
+  topMatrix: Matrix,
+  nbrTopRows: number;
+  nbrLeftColumns: number;
+}
